ChordFinder

Documentation for ChordFinder.

ChordFinder.all_chords

ChordFinder.Chord

ChordFinder.Chord

Base.transpose

Base.transpose

ChordFinder.chords

ChordFinder.trim

ChordFinder.all_chords — Constant

all_chords
Vector containing all chords
all_chords[45]
Chord(["C♯"], Harmony([61, 65, 68]))

ChordFinder.Chord — Method

Chord(s::String)
Get chord based on name (key in all_chords)

ChordFinder.Chord — Method

Chord(n::Vector{String},h::Vector{String})
Short-hand to define Chord from name-vector and harmony vector.

Base.transpose — Method

transpose(c::Chord, semitones)
transpose the chord by the given amount of semitones
Example:
```{julia}
    transpose(Chord(["C"], ["C","E","G"]),1) == Chord(["C#"], ["C#","F","G#"])
```

Base.transpose — Method

transpose(h::Harmony,semitones)
Transpose the harmony by the given amount of semitones

ChordFinder.chords — Method

chords(h::Harmony)
Return chords matching harmony:
If there is a chord matching the harmony, return that (as a 1-element vector).
Else trim the chord to non-redundant pich-classes and return all matching

ChordFinder.trim — Method

trim(h::Harmony)
Remove redundant notes in harmony. It preserves the order.